+++ activerecord/lib/active_record/relation/finder_methods.rb
@@ -301,8 +301,7 @@ module ActiveRecord
     #   Person.exists?
     def exists?(conditions = :none)
       if Base === conditions
-        conditions = conditions.id
-        ActiveSupport::Deprecation.warn(<<-MSG.squish)
+        raise ArgumentError, <<-MSG.squish
           You are passing an instance of ActiveRecord::Base to `exists?`.
           Please pass the id of the object by calling `.id`.
         MSG
@@ -456,11 +455,10 @@ module ActiveRecord
 
         def find_one(id)
           if ActiveRecord::Base === id
-            id = id.id
-            ActiveSupport::Deprecation.warn(<<-MSG.squish)
-            You are passing an instance of ActiveRecord::Base to `find`.
-            Please pass the id of the object by calling `.id`.
-          MSG
+            raise ArgumentError, <<-MSG.squish
+              You are passing an instance of ActiveRecord::Base to `find`.
+              Please pass the id of the object by calling `.id`.
+            MSG
           end
 
           relation = where(primary_key => id)