+++ activerecord/test/cases/finder_test.rb
@@ -118,7 +118,7 @@ class FinderTest < ActiveRecord::TestCase
   end
 
   def test_find_passing_active_record_object_is_deprecated
-    assert_deprecated do
+    assert_raises(ArgumentError) do
       Topic.find(Topic.last)
     end
   end
@@ -167,8 +167,8 @@ class FinderTest < ActiveRecord::TestCase
     assert_equal false, relation.exists?(false)
   end
 
-  def test_exists_passing_active_record_object_is_deprecated
-    assert_deprecated do
+  def test_exists_passing_active_record_object_is_not_permited
+    assert_raises(ArgumentError) do
       Topic.exists?(Topic.new)
     end
   end